#148: Refactoring layout, adding sqlite caching to NCBI query
