+++ bin/delete_data.py
@@ -13,7 +13,8 @@ from analysis_driver.app_logging import AppLogger, logging_default as log_cfg
 from analysis_driver.constants import ELEMENT_PROJECT_ID, ELEMENT_SAMPLE_INTERNAL_ID, ELEMENT_RUN_ELEMENTS,\
     ELEMENT_PROCS, ELEMENT_RUN_NAME, ELEMENT_STATUS, ELEMENT_PROC_ID, ELEMENT_USEABLE, ELEMENT_FASTQS_DELETED,\
     ELEMENT_DELIVERED, ELEMENT_LANE, DATASET_DELETED
-from analysis_driver import rest_communication, executor, clarity, util
+from analysis_driver import executor, util
+from analysis_driver.external_data import rest_communication, clarity
 
 
 class Deleter(AppLogger):
