+++ bin/deliver_reviewed_data.py
@@ -1,15 +1,16 @@
 __author__ = 'tcezard'
-import sys
-import os
 import argparse
-import logging
 import datetime
+import logging
+import os
+import sys
 from collections import defaultdict
 
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 from analysis_driver.config import default as cfg
 from analysis_driver.app_logging import AppLogger, logging_default as log_cfg
-from analysis_driver import executor, clarity, rest_communication
+from analysis_driver import executor, clarity
+from analysis_driver.external_data import rest_communication
 from analysis_driver.exceptions import AnalysisDriverError
